+++ styles/page/_parts.scss
@@ -1,7 +1,7 @@
 // 個別ページ
-@use '../common/variables' as *;
-@use '../common/mixin';
-
+@use "../common/variables" as *;
+@use "../common/mixin";
+@use "sass:color";
 
 .fv_cube {
   height: 100vh;
@@ -15,7 +15,6 @@
     text-align: left;
   }
 }
-
 
 .drawerMenu {
   position: fixed;
@@ -33,7 +32,7 @@
     margin-inline: auto;
     li {
       margin-bottom: 25px;
-      font-family: 'Josefin Sans', sans-serif;
+      font-family: "Josefin Sans", sans-serif;
       font-weight: 200;
       opacity: $opacity;
     }
@@ -73,33 +72,31 @@
 }
 
 .errorPage {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	height: 100vh;
-	text-align: center;
-	background-color: $white;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  text-align: center;
+  background-color: $white;
 
-	h1 {
+  h1 {
     margin-top: 0;
-		margin-bottom: 0.5rem;
-		font-size: 6rem;
-		font-weight: bold;
-	}
+    margin-bottom: 0.5rem;
+    font-size: 6rem;
+    font-weight: bold;
+  }
 
-	h2 {
-		font-size: 2rem;
-		color: darken( rgb(39, 39, 39), 10%);
-	}
+  h2 {
+    font-size: 2rem;
+    color: color.adjust(rgb(39, 39, 39), $lightness: -10%);
+  }
 
-	p {
-		margin-top: 0.5rem;
-		color: lighten( rgb(39, 39, 39), 20%);
-	}
+  p {
+    margin-top: 0.5rem;
+    color: color.adjust(rgb(39, 39, 39), $lightness: 20%);
+  }
 }
-
-
 
 @media (max-width: 1024px) {
   .btn-trigger {
@@ -197,7 +194,6 @@
     }
   }
   .drawerMenu {
-
     position: fixed;
     top: 0;
     z-index: -100;
@@ -213,7 +209,7 @@
       margin-inline: auto;
       li {
         margin-bottom: 25px;
-        font-family: 'Josefin Sans', sans-serif;
+        font-family: "Josefin Sans", sans-serif;
         font-weight: 200;
         opacity: $opacity;
       }
@@ -230,7 +226,6 @@
 }
 
 @include mixin.sp {
-
   .aboutTop_container {
     flex-direction: column;
     gap: rem(10) 0;
