+++ tsconfig.json
@@ -38,7 +38,7 @@
     ".next/types/**/*.ts",
     "app/api/skills/back/router.js",
     "app/api/back/router.js",
-    "next.config.js",
+    "next.config.ts",
     "__tests__/integretion/blog.test.jsx"
 , "__tests__/unit/.test.tsx"  ],
   "exclude": ["node_modules", "**/*.test.tsx", "**/*.storybook.tsx", "app/cube/page.tsx", "dist"]
